LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/access-to-coop-page-from-openee_coop-ro_cross-origin.https-expected.txt


FAIL access-to-coop-page-from-openee, cross-origin assert_equals: expected (string) "coop" but got (undefined) undefined
FAIL access-to-coop-page-from-openee, cross-origin + redirect) assert_equals: expected (string) "coop" but got (undefined) undefined
